+++ middleware/fibermiddleware_test.go
@@ -1,50 +1,23 @@
-/*
- * Copyright 2019 Mia srl
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package glogger
+package middleware
 
 import (
-	"bytes"
-	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
 
+	"github.com/gofiber/fiber/v2"
+	"github.com/mia-platform/glogger/v2"
 	"github.com/sirupsen/logrus"
 	"github.com/sirupsen/logrus/hooks/test"
+	"github.com/valyala/fasthttp"
 	"gotest.tools/assert"
 )
 
-const hostname = "my-host.com"
-const port = "3030"
-const reqIDKey = "reqId"
-const userAgent = "goHttp"
-const bodyBytes = 0
-const path = "/my-req"
-const clientHost = "client-host"
-
-var ip string
-var defaultRequestPath = fmt.Sprintf("http://%s:%s/my-req", hostname, port)
-
-func testMockMuxMiddlewareInvocation(next http.HandlerFunc, requestID string, logger *logrus.Logger, requestPath string) *test.Hook {
+func testMockFiberMiddlewareInvocation(handler fiber.Handler, requestID string, logger *logrus.Logger, requestPath string) *test.Hook {
 	if requestPath == "" {
-		requestPath = defaultRequestPath
+		requestPath = "/my-req"
 	}
 	// create a request
 	req := httptest.NewRequest(http.MethodGet, requestPath, nil)
@@ -63,87 +36,59 @@ func testMockMuxMiddlewareInvocation(next http.HandlerFunc, requestID string, lo
 	if logger != nil {
 		hook = test.NewLocal(logger)
 	}
-	handler := RequestMuxMiddlewareLogger(logger, []string{"/-/"})
-	// invoke the handler
-	server := handler(next)
-	// Create a response writer
-	writer := httptest.NewRecorder()
-	// Serve HTTP server
-	server.ServeHTTP(writer, req)
+
+	// invoke the middleware
+	app := fiber.New()
+	c := app.AcquireCtx(&fasthttp.RequestCtx{})
+	defer app.ReleaseCtx(c)
+	app.Use(RequestFiberMiddlewareLogger(logger, []string{"/-/"}))
+	app.Get(requestPath, handler)
+
+	app.Test(req)
+
 	return hook
 }
 
-func assertJSON(t *testing.T, str string) error {
-	var fields logrus.Fields
-
-	err := json.Unmarshal([]byte(str), &fields)
-	return err
-}
-
-type ExpectedLogFields struct {
-	Level     logrus.Level
-	RequestID string
-	Message   string
-}
-
-type ExpectedIncomingLogFields struct {
-	Method        string
-	Path          string
-	Hostname      string
-	ForwardedHost string
-	Original      string
-	IP            string
-}
-
-type ExpectedOutcomingLogFields struct {
-	Method        string
-	Path          string
-	Hostname      string
-	ForwardedHost string
-	Original      string
-	IP            string
-	Bytes         int
-	StatusCode    int
-}
-
-func TestMuxLogMiddleware(t *testing.T) {
+func TestFiberLogMiddleware(t *testing.T) {
+	hostname := "example.com"
 	t.Run("create a middleware", func(t *testing.T) {
 		called := false
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
 			called = true
-		})
-		testMockMuxMiddlewareInvocation(handler, "", nil, "")
+			return nil
+		}, "", nil, "")
 
 		assert.Assert(t, called, "handler is not called")
 	})
 
-	t.Run("log is a JSON also with trouble getting logger from context", func(t *testing.T) {
-		var buffer bytes.Buffer
-		logger, _ := InitHelper(InitOptions{Level: "trace"})
-		logger.Out = &buffer
-		const logMessage = "New log message"
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			ctx := context.WithValue(r.Context(), loggerKey{}, "notALogger")
-			Get(ctx).Info(logMessage)
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, "", logger, "")
+	// t.Run("log is a JSON also with trouble getting logger from context", func(t *testing.T) {
+	// 	var buffer bytes.Buffer
+	// 	logger, _ := glogger.InitHelper(glogger.InitOptions{Level: "trace"})
+	// 	logger.Out = &buffer
+	// 	const logMessage = "New log message"
+	// 	hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+	// 		ctx := context.WithValue(c.UserContext(), loggerKey{}, "notALogger")
+	// 		Get(ctx).Info(logMessage)
+	// 		return nil
+	// 	}, "", logger, "")
 
-		assert.Equal(t, len(hook.AllEntries()), 2, "Number of logs is not 2")
-		str := buffer.String()
+	// 	assert.Equal(t, len(hook.AllEntries()), 2, "Number of logs is not 2")
+	// 	str := buffer.String()
 
-		for i, value := range strings.Split(strings.TrimSpace(str), "\n") {
-			err := assertJSON(t, value)
-			assert.Equal(t, err, nil, "log %d is not a JSON", i)
-		}
-	})
+	// 	for i, value := range strings.Split(strings.TrimSpace(str), "\n") {
+	// 		err := assertJSON(t, value)
+	// 		assert.Equal(t, err, nil, "log %d is not a JSON", i)
+	// 	}
+	// })
 
 	t.Run("middleware correctly passing configured logger with request id from request header", func(t *testing.T) {
 		const statusCode = 400
 		const requestID = "my-req-id"
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, nil, "")
+
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, requestID, nil, "")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 2, "Unexpected entries length.")
@@ -190,11 +135,12 @@ func TestMuxLogMiddleware(t *testing.T) {
 	t.Run("passing a content-length header by default", func(t *testing.T) {
 		const statusCode = 200
 		const requestID = "my-req-id"
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-			w.Header().Set("content-length", "10")
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, nil, "")
+
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			c.Set("content-length", "10")
+			return nil
+		}, requestID, nil, "")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 2, "Unexpected entries length.")
@@ -242,11 +188,12 @@ func TestMuxLogMiddleware(t *testing.T) {
 		const statusCode = 200
 		const requestID = "my-req-id"
 		contentToWrite := []byte("testing\n")
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-			w.Write(contentToWrite)
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, nil, "")
+
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			c.Write(contentToWrite)
+			return nil
+		}, requestID, nil, "")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 2, "Unexpected entries length.")
@@ -293,11 +240,12 @@ func TestMuxLogMiddleware(t *testing.T) {
 	t.Run("using info level returning only outcomingRequest", func(t *testing.T) {
 		const statusCode = 200
 		const requestID = "my-req-id"
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
+
 		logger, _ := test.NewNullLogger()
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, logger, "")
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, requestID, logger, "")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 1, "Unexpected entries length.")
@@ -328,11 +276,11 @@ func TestMuxLogMiddleware(t *testing.T) {
 		const requestID = "my-req-id"
 		var requestPathWithoutPort = fmt.Sprintf("http://%s/my-req", hostname)
 
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
 		logger, _ := test.NewNullLogger()
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, logger, requestPathWithoutPort)
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, requestID, logger, requestPathWithoutPort)
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 1, "Unexpected entries length.")
@@ -362,15 +310,14 @@ func TestMuxLogMiddleware(t *testing.T) {
 		const statusCode = 200
 		const requestID = "my-req-id"
 		const pathWithQuery = "/my-req?foo=bar&some=other"
-		var requestPathWithoutPort = fmt.Sprintf("http://%s%s", hostname, pathWithQuery)
 
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
-		logger, _ := InitHelper(InitOptions{
+		logger, _ := glogger.InitHelper(glogger.InitOptions{
 			DisableHTMLEscape: true,
 		})
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, logger, requestPathWithoutPort)
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, requestID, logger, pathWithQuery)
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 1, "Unexpected entries length.")
@@ -381,13 +328,14 @@ func TestMuxLogMiddleware(t *testing.T) {
 		hook.Reset()
 	})
 
-	t.Run("middleware correctly passing configured logger with request id from request header", func(t *testing.T) {
+	t.Run("do not log skipped paths", func(t *testing.T) {
 		const statusCode = 200
 		const requestID = "my-req-id"
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, requestID, nil, "/-/healthz")
+
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, requestID, nil, "/-/healthz")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 0, "Unexpected entries length.")
@@ -397,10 +345,11 @@ func TestMuxLogMiddleware(t *testing.T) {
 
 	t.Run("middleware correctly create request id if not present in header", func(t *testing.T) {
 		const statusCode = 400
-		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.WriteHeader(statusCode)
-		})
-		hook := testMockMuxMiddlewareInvocation(handler, "", nil, "")
+
+		hook := testMockFiberMiddlewareInvocation(func(c *fiber.Ctx) error {
+			c.Status(statusCode)
+			return nil
+		}, "", nil, "")
 
 		entries := hook.AllEntries()
 		assert.Equal(t, len(entries), 2, "Unexpected entries length.")
@@ -411,6 +360,7 @@ func TestMuxLogMiddleware(t *testing.T) {
 			Level:   logrus.TraceLevel,
 			Message: "incoming request",
 		})
+		assert.Assert(t, incomingRequestID != "")
 		incomingRequestAssertions(t, incomingRequest, ExpectedIncomingLogFields{
 			Method:        http.MethodGet,
 			Path:          path,
@@ -441,50 +391,4 @@ func TestMuxLogMiddleware(t *testing.T) {
 
 		hook.Reset()
 	})
-}
-
-func logAssertions(t *testing.T, logEntry *logrus.Entry, expected ExpectedLogFields) string {
-	t.Helper()
-	assert.Equal(t, logEntry.Level, expected.Level, "Unexpected level of log for log in incoming request")
-	assert.Equal(t, logEntry.Message, expected.Message, "Unexpected message of log for log in incoming request")
-	requestID := logEntry.Data[reqIDKey]
-	_, ok := requestID.(string)
-	assert.Assert(t, ok, "Unexpected or empty requestID for log in incoming request")
-	if expected.RequestID != "" {
-		assert.Equal(t, requestID, expected.RequestID, "Unexpected requestID for log in incoming request")
-	}
-	return requestID.(string)
-}
-
-func incomingRequestAssertions(t *testing.T, incomingRequestLogEntry *logrus.Entry, expected ExpectedIncomingLogFields) {
-	http := incomingRequestLogEntry.Data["http"].(HTTP)
-	assert.Equal(t, http.Request.Method, expected.Method, "Unexpected http method for log in incoming request")
-	assert.Equal(t, http.Request.UserAgent["original"], expected.Original, "Unexpected original userAgent for log of request completed")
-
-	url := incomingRequestLogEntry.Data["url"].(URL)
-	assert.Equal(t, url.Path, expected.Path, "Unexpected http uri path for log in incoming request")
-
-	host := incomingRequestLogEntry.Data["host"].(Host)
-	assert.Equal(t, host.Hostname, expected.Hostname, "Unexpected hostname for log of request completed")
-	assert.Equal(t, host.ForwardedHost, expected.ForwardedHost, "Unexpected forwaded hostname for log of request completed")
-	assert.Equal(t, host.IP, expected.IP, "Unexpected ip for log of request completed")
-}
-
-func outcomingRequestAssertions(t *testing.T, outcomingRequestLogEntry *logrus.Entry, expected ExpectedOutcomingLogFields) {
-	http := outcomingRequestLogEntry.Data["http"].(HTTP)
-	assert.Equal(t, http.Request.Method, expected.Method, "Unexpected http method for log in incoming request")
-	assert.Equal(t, http.Request.UserAgent["original"], expected.Original, "Unexpected original userAgent for log of request completed")
-	assert.Equal(t, http.Response.StatusCode, expected.StatusCode, "Unexpected status code for log of request completed")
-	assert.Equal(t, http.Response.Body["bytes"], expected.Bytes, "Unexpected status code for log of request completed")
-
-	url := outcomingRequestLogEntry.Data["url"].(URL)
-	assert.Equal(t, url.Path, expected.Path, "Unexpected http uri path for log in incoming request")
-
-	host := outcomingRequestLogEntry.Data["host"].(Host)
-	assert.Equal(t, host.Hostname, expected.Hostname, "Unexpected hostname for log of request completed")
-	assert.Equal(t, host.ForwardedHost, expected.ForwardedHost, "Unexpected forwaded hostname for log of request completed")
-	assert.Equal(t, host.IP, expected.IP, "Unexpected ip for log of request completed")
-
-	_, ok := outcomingRequestLogEntry.Data["responseTime"].(float64)
-	assert.Assert(t, ok, "Invalid took duration for log of request completed")
 }
