+++ middleware/readableresponsewriter_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package glogger
+package middleware
 
 import (
 	"net/http"
